fix(activerecord): rename src/railties → src/trailties for global alias consistency
Copilot review #1 caught that the new global path-segment alias
(`railties` → `trailties`) breaks the api:compare match for
`packages/activerecord/src/railties/{controller,job}-runtime.ts` —
their Rails sources live under `activerecord/lib/active_record/railties/`,
which under the new alias maps to `trailties/`. Rename the AR directory
to match (4 file moves + 2 import updates) so the unified trails naming
convention covers existing files too, not just the new trailties package.
